added possibility in swagger to authorize jwt
Needed an authorize button in order to access certain endpoints. So, now, Swagger can take a token and it will automatically send the token as a header for it to be further used.
Example use case:
Subs/SubscribeToParty/id => subscribes the currently logged in user to the party with partyId id, further creating a link between the id of the party and the id of the user in the Subscriptions table.